Add remote modules support for OTB nixpkgs (#15)
* [ADD] Add support for multiple new OTB remote modules

* Enabled optional integration of various OTB remote modules like MLUtils, Prefetch, and Phenology.

* [ADD] Add otb-dev and otb-dev-docker packages

* Add new `otb-dev` package build options for all the remote modules.

 * Also added `otb-dev-docker` for image based on all the remote modules.

* [ADD] Add OTB_APPLICATION_PATH to environment variables

* Add OTB_APPLICATION_PATH so that applications are importable via native python api of OTB

* Additionally add a USER as `otbuser` and HOME directory for docker images

* [FIX] Update the Makefile command for the updated docker package

* [CI] Update CI build commands to include both the builds for otb modules

* Builds OTB with all the remote modules and also build with minimal modules

* [REFAC] Make all remote modules to be not activated by default

* [DOC] Update Docker build instructions and handle remote modules

* Add documentation for examples on remote modules,

* Add some note on OTBTF related to TF library

* [FIX] Fix typos in OTB default options

* [FIX] Fix the wrong package name in CI build for OTB

* [DOC] Correct typo in README.md

* Add aarch64 runner (#16)

* [CI] Add AArch64 GitHub Action Workflow

* Add a new GitHub Actions workflow for AArch64 builds using ArchLinux example from https://github.com/marketplace/actions/arm-runner. This prepares an ArchLinux image, sets up the nix environment, and builds the project for AArch64.

* [FIX] Fix Aarch64 workflow build system

* Fix the matrix.machine.platform with aarch64-linux in Nix build commands.

* [FIX] Fix build steps in aarch64 workflow

* Move the 'Build system' task to correctly follow the cache setup.

* [CI] Add permissions to CI workflows and update build steps

* Updated the CI workflows to include `contents: read` and `id-token: write` permissions. Also adjusted build commands to focus on `otb-dev` target and included the use of `magic-nix-cache-action`.

* [CI] Remove deprecated aarch64 GitHub Actions workflow

* The aarch64 GitHub Actions workflow file (.github/workflows/aarch64.yml) was deleted.

* This action is no more needed as the ci.yml works fine with Nix-Magic-Cache

* Rather than building 2 derivation i.e. otb-dev (all modules) and otb (minimal modules) which takes a lot of time for ARM64 build and gets timed out so we build only one derivation i.e otb-dev which contains all the modules.

* [REFAC] Move remote module to their own separate nix file for better readability

* Move the remote module to its own separate file for better readability, also add information regarding remote module code LICENSE and copyright information

* Updated related configurations in `flake.nix` and `docker.nix` to include licensing information and appropriate fetchGit instructions.

+++ docker.nix
@@ -1,3 +1,17 @@
+#   Copyright 2024 Pratyush Das
+#
+#   Licensed under the Apache License, Version 2.0 (the "License");
+#   you may not use this file except in compliance with the License.
+#   You may obtain a copy of the License at
+#
+#       http://www.apache.org/licenses/LICENSE-2.0
+#
+#   Unless required by applicable law or agreed to in writing, software
+#   distributed under the License is distributed on an "AS IS" BASIS,
+#   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+#   See the License for the specific language governing permissions and
+#   limitations under the License.
+#   https://github.com/efabless/openlane2/blob/main/nix/docker.nix
 {
   nix2containerPkgs,
   python3,
@@ -9,6 +23,8 @@
   pkgs,
   ...
 }: let
+  user = "otbuser";
+  home-dir = "/home/${user}";
   py-env = (
     python3.withPackages (pp:
       with pp;
@@ -24,7 +40,7 @@ in
   nix2containerPkgs.nix2container.buildImage rec {
     name = img-name;
     tag = img-tag;
-
+    initializeNixDatabase = true;
     copyToRoot = pkgs.buildEnv {
       name = "root";
       paths = with pkgs.dockerTools;
@@ -50,6 +66,7 @@ in
           git
           neovim-unwrapped
           zsh
+          nix
 
           # Library
           otb
@@ -59,6 +76,12 @@ in
 
       pathsToLink = ["/bin" "/etc" "/var" "/run" "/tmp" "/lib"];
       postBuild = ''
+        if -e $HOME; then
+            echo "home directory exists"
+            exit 1
+        fi
+            mkdir -p $out/${home-dir}
+        mkdir -p ~/.config/nix/ && printf "experimental-features = nix-command flakes\n" > ~/.config/nix/nix.conf
         mkdir -p $out/tmp
         mkdir -p $out/etc
         cat <<HEREDOC > $out/etc/zshrc
@@ -79,8 +102,10 @@ in
         "EDITOR=nvim"
         "PYTHONPATH=${py-env-sitepackages}:${otbLibPath}/otb/python"
         "PATH=${py-env-bin}:${otbBinPath}:/bin"
+        "OTB_APPLICATION_PATH=${otbLibPath}/otb/applications"
         "TMPDIR=/tmp"
+        "USER=${user}"
+        "HOME=${home-dir}"
       ];
     };
   }
-
